Registration Working - all other database connctions disconnected for now
