+++ package.json
@@ -20,6 +20,7 @@
     "nodemon": "^1.18.7"
   },
   "dependencies": {
+    "bcrypt": "^3.0.5",
     "cloudinary": "^1.14.0",
     "dotenv": "^7.0.0",
     "express": "^4.16.3",
